chore(test-tags): refactor test tags imports according to folder reorganization
