+++ app/src/androidTest/java/com/android/unio/components/BottomNavigationTest.kt
@@ -10,7 +10,7 @@ import com.android.unio.model.event.EventViewModel
 import com.android.unio.model.image.ImageRepositoryFirebaseStorage
 import com.android.unio.model.search.SearchRepository
 import com.android.unio.model.search.SearchViewModel
-import com.android.unio.model.strings.test_tags.NavigationActionTestTags
+import com.android.unio.model.strings.test_tags.navigation.NavigationActionTestTags
 import com.android.unio.model.user.UserRepository
 import com.android.unio.model.user.UserRepositoryFirestore
 import com.android.unio.model.user.UserViewModel
