+++ app/src/androidTest/java/com/android/unio/components/ScreenDisplayingTest.kt
@@ -24,20 +24,20 @@ import com.android.unio.model.map.nominatim.NominatimLocationRepository
 import com.android.unio.model.map.nominatim.NominatimLocationSearchViewModel
 import com.android.unio.model.search.SearchRepository
 import com.android.unio.model.search.SearchViewModel
-import com.android.unio.model.strings.test_tags.AccountDetailsTestTags
-import com.android.unio.model.strings.test_tags.AssociationProfileTestTags
-import com.android.unio.model.strings.test_tags.EmailVerificationTestTags
-import com.android.unio.model.strings.test_tags.EventCreationTestTags
-import com.android.unio.model.strings.test_tags.EventDetailsTestTags
-import com.android.unio.model.strings.test_tags.ExploreContentTestTags
-import com.android.unio.model.strings.test_tags.ExploreTestTags
-import com.android.unio.model.strings.test_tags.HomeTestTags
-import com.android.unio.model.strings.test_tags.MapTestTags
-import com.android.unio.model.strings.test_tags.SavedTestTags
-import com.android.unio.model.strings.test_tags.SettingsTestTags
-import com.android.unio.model.strings.test_tags.SomeoneElseUserProfileTestTags
-import com.android.unio.model.strings.test_tags.UserProfileTestTags
-import com.android.unio.model.strings.test_tags.WelcomeTestTags
+import com.android.unio.model.strings.test_tags.authentication.AccountDetailsTestTags
+import com.android.unio.model.strings.test_tags.association.AssociationProfileTestTags
+import com.android.unio.model.strings.test_tags.authentication.EmailVerificationTestTags
+import com.android.unio.model.strings.test_tags.event.EventCreationTestTags
+import com.android.unio.model.strings.test_tags.event.EventDetailsTestTags
+import com.android.unio.model.strings.test_tags.explore.ExploreContentTestTags
+import com.android.unio.model.strings.test_tags.explore.ExploreTestTags
+import com.android.unio.model.strings.test_tags.home.HomeTestTags
+import com.android.unio.model.strings.test_tags.map.MapTestTags
+import com.android.unio.model.strings.test_tags.saved.SavedTestTags
+import com.android.unio.model.strings.test_tags.settings.SettingsTestTags
+import com.android.unio.model.strings.test_tags.user.SomeoneElseUserProfileTestTags
+import com.android.unio.model.strings.test_tags.user.UserProfileTestTags
+import com.android.unio.model.strings.test_tags.authentication.WelcomeTestTags
 import com.android.unio.model.user.User
 import com.android.unio.model.user.UserRepositoryFirestore
 import com.android.unio.model.user.UserViewModel
