+++ app/src/androidTest/java/com/android/unio/components/authentication/overlay/InterestOverlayTest.kt
@@ -8,7 +8,7 @@ import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performScrollTo
 import com.android.unio.TearDown
-import com.android.unio.model.strings.test_tags.InterestsOverlayTestTags
+import com.android.unio.model.strings.test_tags.authentication.InterestsOverlayTestTags
 import com.android.unio.model.user.Interest
 import com.android.unio.ui.authentication.overlay.InterestOverlay
 import org.junit.Before
