+++ app/src/androidTest/java/com/android/unio/components/event/EventCardTest.kt
@@ -21,7 +21,7 @@ import com.android.unio.model.event.EventType
 import com.android.unio.model.event.EventViewModel
 import com.android.unio.model.image.ImageRepositoryFirebaseStorage
 import com.android.unio.model.notification.NotificationWorker
-import com.android.unio.model.strings.test_tags.EventCardTestTags
+import com.android.unio.model.strings.test_tags.event.EventCardTestTags
 import com.android.unio.model.user.UserRepositoryFirestore
 import com.android.unio.model.user.UserViewModel
 import com.android.unio.ui.event.EventCard
