+++ app/src/androidTest/java/com/android/unio/components/event/EventCreationTest.kt
@@ -28,8 +28,8 @@ import com.android.unio.model.map.nominatim.NominatimLocationSearchViewModel
 import com.android.unio.model.search.SearchRepository
 import com.android.unio.model.search.SearchViewModel
 import com.android.unio.model.strings.TextLengthSamples
-import com.android.unio.model.strings.test_tags.EventCreationOverlayTestTags
-import com.android.unio.model.strings.test_tags.EventCreationTestTags
+import com.android.unio.model.strings.test_tags.event.EventCreationOverlayTestTags
+import com.android.unio.model.strings.test_tags.event.EventCreationTestTags
 import com.android.unio.ui.event.EventCreationScreen
 import com.android.unio.ui.navigation.NavigationAction
 import com.google.firebase.Firebase
