+++ app/src/androidTest/java/com/android/unio/components/map/MapScreenTest.kt
@@ -17,7 +17,7 @@ import com.android.unio.model.event.EventRepositoryFirestore
 import com.android.unio.model.event.EventViewModel
 import com.android.unio.model.image.ImageRepositoryFirebaseStorage
 import com.android.unio.model.map.MapViewModel
-import com.android.unio.model.strings.test_tags.MapTestTags
+import com.android.unio.model.strings.test_tags.map.MapTestTags
 import com.android.unio.model.user.User
 import com.android.unio.model.user.UserRepositoryFirestore
 import com.android.unio.model.user.UserViewModel
