+++ app/src/androidTest/java/com/android/unio/end2end/SearchTest.kt
@@ -9,12 +9,12 @@ import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performTextInput
 import androidx.test.filters.LargeTest
-import com.android.unio.model.strings.test_tags.AssociationProfileTestTags
-import com.android.unio.model.strings.test_tags.BottomNavBarTestTags
-import com.android.unio.model.strings.test_tags.EventCardTestTags
-import com.android.unio.model.strings.test_tags.EventDetailsTestTags
-import com.android.unio.model.strings.test_tags.ExploreContentTestTags
-import com.android.unio.model.strings.test_tags.HomeTestTags
+import com.android.unio.model.strings.test_tags.association.AssociationProfileTestTags
+import com.android.unio.model.strings.test_tags.navigation.BottomNavBarTestTags
+import com.android.unio.model.strings.test_tags.event.EventCardTestTags
+import com.android.unio.model.strings.test_tags.event.EventDetailsTestTags
+import com.android.unio.model.strings.test_tags.explore.ExploreContentTestTags
+import com.android.unio.model.strings.test_tags.home.HomeTestTags
 import com.google.firebase.Firebase
 import com.google.firebase.auth.auth
 import dagger.hilt.android.testing.HiltAndroidTest
