+++ app/src/main/java/com/android/unio/ui/authentication/AccountDetails.kt
@@ -39,7 +39,7 @@ import com.android.unio.model.event.Event
 import com.android.unio.model.firestore.emptyFirestoreReferenceList
 import com.android.unio.model.image.ImageViewModel
 import com.android.unio.model.strings.StoragePathsStrings
-import com.android.unio.model.strings.test_tags.AccountDetailsTestTags
+import com.android.unio.model.strings.test_tags.authentication.AccountDetailsTestTags
 import com.android.unio.model.user.AccountDetailsError
 import com.android.unio.model.user.Interest
 import com.android.unio.model.user.User
