+++ app/src/main/java/com/android/unio/ui/authentication/EmailVerification.kt
@@ -31,7 +31,7 @@ import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import com.android.unio.R
-import com.android.unio.model.strings.test_tags.EmailVerificationTestTags
+import com.android.unio.model.strings.test_tags.authentication.EmailVerificationTestTags
 import com.android.unio.model.user.UserViewModel
 import com.android.unio.ui.navigation.NavigationAction
 import com.android.unio.ui.navigation.Screen
