+++ app/src/main/java/com/android/unio/ui/authentication/Welcome.kt
@@ -42,7 +42,7 @@ import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.text.style.TextAlign.Companion.Center
 import androidx.compose.ui.unit.dp
 import com.android.unio.R
-import com.android.unio.model.strings.test_tags.WelcomeTestTags
+import com.android.unio.model.strings.test_tags.authentication.WelcomeTestTags
 import com.android.unio.model.user.SignInState
 import com.android.unio.model.user.UserViewModel
 import com.android.unio.model.user.isValidEmail
