+++ app/src/main/java/com/android/unio/ui/event/overlay/AssociationsOverlay.kt
@@ -41,7 +41,7 @@ import androidx.wear.compose.material.Text
 import com.android.unio.R
 import com.android.unio.model.association.Association
 import com.android.unio.model.search.SearchViewModel
-import com.android.unio.model.strings.test_tags.EventCreationOverlayTestTags
+import com.android.unio.model.strings.test_tags.event.EventCreationOverlayTestTags
 import com.android.unio.ui.theme.AppTypography
 
 /**
